re-organizing files slightly, Fernet is the goat :3
